Separate event listener registration from dispatcher
